Extract class completed
