+++ src/main/java/edu/mit/blocks/codeblocks/BlockB.java
@@ -1,0 +1,2 @@
+package edu.mit.blocks.codeblocks;public class BlockB {
+}
