+++ src/main/java/edu/mit/blocks/renderable/CommentNode.java
@@ -1,0 +1,13 @@
+package edu.mit.blocks.renderable;
+
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.Node;
+import org.w3c.dom.Text;
+
+import javax.swing.*;
+
+public class CommentNode extends JPanel {
+
+
+}
