+++ src/main/java/edu/mit/blocks/renderable/RenderableBlockAccomodate.java
@@ -1,0 +1,5 @@
+package edu.mit.blocks.renderable;
+
+public class RenderableBlockAccomodate {
+
+}
